backend/requirements.txt
django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pillow>=10.0.0
